B alaa

+++ lib/view/screens/profile_screens/screens/delete_user.dart
@@ -17,21 +17,21 @@ class DeleteUser extends StatefulWidget {
 }
 
 class _DeleteUser extends State<DeleteUser> {
-
-  UserProvider userProvider=UserProvider();
+  UserProvider userProvider = UserProvider();
 
   @override
   Widget build(BuildContext context) {
-      print("FirebaseAuth.instance.currentUser${FirebaseAuth.instance.currentUser?.email}");
-      print("object${FirebaseAuth.instance.currentUser?.uid}");
-    
+    print(
+        "FirebaseAuth.instance.currentUser${FirebaseAuth.instance.currentUser?.email}");
+    print("object${FirebaseAuth.instance.currentUser?.uid}");
+
     return Scaffold(
       backgroundColor: ThemeManager.background,
       appBar: AppBar(
         centerTitle: true,
         backgroundColor: ThemeManager.background,
         title: Text(
-          'Delete User',
+          'Delete Account',
           style: ThemeManager.textStyle.copyWith(
             fontSize: LayoutManager.widthNHeight0(context, 1) * 0.05,
             fontWeight: FontWeight.bold,
@@ -64,12 +64,52 @@ class _DeleteUser extends State<DeleteUser> {
                       Divider(height: 1, color: Colors.grey[300]),
                       InkWell(
                         onTap: () async {
-               
+                          showDialog(
+                            context: context,
+                            builder: (BuildContext context) {
+                              return AlertDialog(
+                                backgroundColor: ThemeManager.primary,
+                                title: Text(
+                                  'Confirm Deletion',
+                                  style: TextStyle(color: Colors.white),
+                                ),
+                                content: Text(
+                                  // textAlign: TextAlign.start,
+                                  'Are you sure you want to delete your account ?',
+                                  style: TextStyle(color: Colors.white,fontSize: 16),
+                                ),
+                                actions: [
+                                  TextButton(
+                                    onPressed: () {
+                                      Navigator.of(context).pop();
+                                    },
+                                    child: Text(
+                                      'Cancel',
+                                      style: TextStyle(
+                                          color:ThemeManager.second,
+                                          fontWeight: FontWeight.bold),
+                                    ),
+                                  ),
+                                  TextButton(
+                                    onPressed: () {
+                                      userProvider.deleteUserprovider();
 
-                          userProvider.deleteUserprovider();
-
-                          //log out
-                          Navigator.of(context).pushReplacementNamed(signIn);
+                                      //log out
+                                      Navigator.of(context)
+                                          .pushReplacementNamed(signIn);
+                                    },
+                                    child: Text(
+                                      'Delete',
+                                      style: TextStyle(
+                                          color:
+                                              Color.fromARGB(255, 255, 32, 16),
+                                          fontWeight: FontWeight.bold),
+                                    ),
+                                  ),
+                                ],
+                              );
+                            },
+                          );
                         },
                         child: Center(
                           child: Container(
@@ -83,7 +123,7 @@ class _DeleteUser extends State<DeleteUser> {
                             ),
                             child: Center(
                               child: Text(
-                                "Delete User",
+                                "Delete My Account",
                                 style: TextStyle(
                                   color: ThemeManager.second,
                                   fontSize: 20,
